DEV-1188 Improve handling of unknown service host
The host of a service is defined by the name of said service in the
sense that the name is used to look up the host in the config.yml.
If that field, based on the service name, is not found in the config
file, it logged a warning that wasn't very useful. This has been
expanded so that it also logs the service name.

Also fixed the host config for "mediahaven-api".