DEV-1188 Rename TestUnknownHostService
Pytest assumes that a class starting with the name 'Test' is a test
class and tries to collect the tests in that class. Pytest threw a
warning that it was not possible to collect the tests because that
class had an __init__ function. However, `TestUnknownHostService` is
just a subclass of `Service` that should only be used in tests and
doesn't actually contain any tests.