DEV-1215 Replace auto-ack mechanism
When using the auto-ack mechanism, a message is considered to be
successfully delivered immediately after it is sent. Even if the
processing of the message would fail, the message will be acked.

We'll send out a manual ack after the message has been successfully
processed. In the case an error happens, a nack will be send. This also
allows for dead-letter functionality.