DEV-1297 Cache CP name
In the case of a create event, the Organisations API will be queried to
return the CP MAM Name for a given OR-ID. This CP MAM Name won't change
often and it will also be a small set of OR-IDs that will be potentially
queried. Sending out this call for every event is bit wasteful. We will
now cache this information after a successful call to the Organisations
API.