DEV-1574 Add MD5 check if item is collateral
It is possible to get the exact same subtitles multiple times but with
a different object key. The check if an item is already ingested based on
the md5 was removed for VRT, regardless of the type of item. This means
that the same subtitle would be ingested and linked to the fragment
multiple times.

Add a MD5 check if an item is a collateral regardless of the CP.