DEV-2321 Fix metadata update in case of collateral
In the case of a collateral, the metadata of the original essence should
be updated with the relation of the collateral. However, there was an issue
with the metadata update call.

The payload of the update metadata call needs to be a record update object.
Only the sidecar was given as an XML payload in the body. Instead of passing
an XML in the body, change it to multipart/form-data and add the missing
parameters as other form-data together with the sidecar XML.