DEV-2509 Fix crash when MAM label is None
It is possible that the result of the org API call has no MAM label. This
is the case when an organisation is not a content partner. Ideally, there
should be no events for these organisations but at the moment, handling
such an event results in a crash.

Handle such an event properly by throwing a NackException.

Also refactor the method:
 - Use return to jump out of the method
 - Remove superfluous exception
 - Decrease try-except blocks

This results in less nesting and improves readability.